pyproject.toml
[project]
name = "louisjohn26"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.5",
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "requests>=2.33.1",
    "rich>=15.0.0",
    "uvicorn>=0.44.0",
]
